Update client ViewPorts when adding NeutralBlubs
Fix a bug where 'ViewPort's were not immediately updated when adding a
'NeutralBlub'.

When adding a 'NeutralBlub', the new 'ViewPort' that was sent out to
clients accidentally discarded the new 'World' with the 'NeutralBlub' in
it. So the 'NeutralBlub' was only visible in the next update where
'addNeutral' was *not* called.

Ooops. This patch updates the clients' 'ViewPort's immediately after
adding a 'NeutralBlub'.

Signed-off-by: Alexander Berntsen <alexander@plaimi.net>
Signed-off-by: Stian Ellingsen    <stian@plaimi.net>
Reviewed-by:   Stian Ellingsen    <stian@plaimi.net>